+++ proofs/adapter/filesystem.php
@@ -21,10 +21,12 @@ return static function() {
             $path = \sys_get_temp_dir().'/innmind/filesystem/';
             (new FS)->remove($path);
 
-            return Filesystem::mount(Path::of($path))->withCaseSensitivity(match (\PHP_OS) {
-                'Darwin' => CaseSensitivity::insensitive,
-                default => CaseSensitivity::sensitive,
-            });
+            return Filesystem::mount(Path::of($path))
+                ->unwrap()
+                ->withCaseSensitivity(match (\PHP_OS) {
+                    'Darwin' => CaseSensitivity::insensitive,
+                    default => CaseSensitivity::sensitive,
+                });
         }),
     );
 
@@ -35,10 +37,12 @@ return static function() {
                 $path = \sys_get_temp_dir().'/innmind/filesystem/';
                 (new FS)->remove($path);
 
-                return Filesystem::mount(Path::of($path))->withCaseSensitivity(match (\PHP_OS) {
-                    'Darwin' => CaseSensitivity::insensitive,
-                    default => CaseSensitivity::sensitive,
-                });
+                return Filesystem::mount(Path::of($path))
+                    ->unwrap()
+                    ->withCaseSensitivity(match (\PHP_OS) {
+                        'Darwin' => CaseSensitivity::insensitive,
+                        default => CaseSensitivity::sensitive,
+                    });
             }),
         )->named('Filesystem');
     }
@@ -55,10 +59,12 @@ return static function() {
 
             $path = \sys_get_temp_dir().'/innmind/filesystem/';
             (new FS)->remove($path);
-            $adapter = Filesystem::mount(Path::of($path))->withCaseSensitivity(match (\PHP_OS) {
-                'Darwin' => CaseSensitivity::insensitive,
-                default => CaseSensitivity::sensitive,
-            });
+            $adapter = Filesystem::mount(Path::of($path))
+                ->unwrap()
+                ->withCaseSensitivity(match (\PHP_OS) {
+                    'Darwin' => CaseSensitivity::insensitive,
+                    default => CaseSensitivity::sensitive,
+                });
 
             $property->ensureHeldBy($assert, $adapter);