+++ tests/Adapter/FilesystemTest.php
@@ -41,7 +41,7 @@ class FilesystemTest extends TestCase
 
     public function testInterface()
     {
-        $adapter = Filesystem::mount(Path::of('/tmp/'));
+        $adapter = Filesystem::mount(Path::of('/tmp/'))->unwrap();
 
         $this->assertInstanceOf(Adapter::class, $adapter);
         $this->assertFalse($adapter->contains(Name::of('foo')));
@@ -66,13 +66,14 @@ class FilesystemTest extends TestCase
         $this->expectException(PathDoesntRepresentADirectory::class);
         $this->expectExceptionMessage('path/to/somewhere');
 
-        Filesystem::mount(Path::of('path/to/somewhere'));
+        Filesystem::mount(Path::of('path/to/somewhere'))->unwrap();
     }
 
     public function testReturnNothingWhenGettingUnknownFile()
     {
         $this->assertNull(
             Filesystem::mount(Path::of('/tmp/'))
+                ->unwrap()
                 ->get(Name::of('foo'))
                 ->match(
                     static fn($file) => $file,
@@ -86,6 +87,7 @@ class FilesystemTest extends TestCase
         $this->assertInstanceOf(
             SideEffect::class,
             Filesystem::mount(Path::of('/tmp/'))
+                ->unwrap()
                 ->remove(Name::of('foo'))
                 ->unwrap(),
         );
@@ -93,7 +95,7 @@ class FilesystemTest extends TestCase
 
     public function testCreateNestedStructure()
     {
-        $adapter = Filesystem::mount(Path::of('/tmp/'));
+        $adapter = Filesystem::mount(Path::of('/tmp/'))->unwrap();
 
         $directory = Directory::of(Name::of('foo'))
             ->add(File::of(Name::of('foo.md'), Content::ofString('# Foo')))
@@ -126,7 +128,7 @@ class FilesystemTest extends TestCase
                 ),
         );
 
-        $adapter = Filesystem::mount(Path::of('/tmp/'));
+        $adapter = Filesystem::mount(Path::of('/tmp/'))->unwrap();
         $this->assertTrue($adapter->contains(Name::of('foo')));
         $this->assertSame(
             '# Foo',
@@ -165,7 +167,7 @@ class FilesystemTest extends TestCase
 
     public function testRemoveFileWhenRemovedFromFolder()
     {
-        $a = Filesystem::mount(Path::of('/tmp/'));
+        $a = Filesystem::mount(Path::of('/tmp/'))->unwrap();
 
         $d = Directory::of(Name::of('foo'));
         $d = $d->add(File::of(Name::of('bar'), Content::ofString('some content')));
@@ -173,7 +175,7 @@ class FilesystemTest extends TestCase
         $d = $d->remove(Name::of('bar'));
         $a->add($d)->unwrap();
         $this->assertSame(1, $d->removed()->size());
-        $a = Filesystem::mount(Path::of('/tmp/'));
+        $a = Filesystem::mount(Path::of('/tmp/'))->unwrap();
         $this->assertFalse(
             $a->get(Name::of('foo'))->match(
                 static fn($dir) => $dir->contains(Name::of('bar')),
@@ -187,7 +189,7 @@ class FilesystemTest extends TestCase
 
     public function testDoesntFailWhenAddindSameDirectoryTwiceThatContainsARemovedFile()
     {
-        $a = Filesystem::mount(Path::of('/tmp/'));
+        $a = Filesystem::mount(Path::of('/tmp/'))->unwrap();
 
         $d = Directory::of(Name::of('foo'));
         $d = $d->add(File::of(Name::of('bar'), Content::ofString('some content')));
@@ -196,7 +198,7 @@ class FilesystemTest extends TestCase
         $a->add($d)->unwrap();
         $a->add($d)->unwrap();
         $this->assertSame(1, $d->removed()->size());
-        $a = Filesystem::mount(Path::of('/tmp/'));
+        $a = Filesystem::mount(Path::of('/tmp/'))->unwrap();
         $this->assertFalse(
             $a->get(Name::of('foo'))->match(
                 static fn($dir) => $dir->contains(Name::of('bar')),
@@ -210,7 +212,7 @@ class FilesystemTest extends TestCase
 
     public function testLoadWithMediaType()
     {
-        $a = Filesystem::mount(Path::of('/tmp/'));
+        $a = Filesystem::mount(Path::of('/tmp/'))->unwrap();
         \file_put_contents(
             '/tmp/some_content.html',
             '<!DOCTYPE html><html><body><answer value="42"/></body></html>',
@@ -233,7 +235,7 @@ class FilesystemTest extends TestCase
 
     public function testRoot()
     {
-        $adapter = Filesystem::mount(Path::of('/tmp/test/'));
+        $adapter = Filesystem::mount(Path::of('/tmp/test/'))->unwrap();
         $adapter
             ->add(File::of(
                 Name::of('foo'),
@@ -294,7 +296,7 @@ class FilesystemTest extends TestCase
 
     public function testAddingTheSameFileTwiceDoesNothing()
     {
-        $adapter = Filesystem::mount(Path::of('/tmp/'));
+        $adapter = Filesystem::mount(Path::of('/tmp/'))->unwrap();
         $file = File::of(
             Name::of('foo'),
             Content::ofString('foo'),
@@ -323,7 +325,7 @@ class FilesystemTest extends TestCase
         $path = \sys_get_temp_dir().'/innmind/filesystem/';
         (new FS)->remove($path);
 
-        $filesystem = Filesystem::mount(Path::of($path));
+        $filesystem = Filesystem::mount(Path::of($path))->unwrap();
 
         $this->expectException(\RuntimeException::class);
         $this->expectExceptionMessage('Path too long');
@@ -363,7 +365,7 @@ class FilesystemTest extends TestCase
                 $path = \sys_get_temp_dir().'/innmind/filesystem/';
                 (new FS)->remove($path);
 
-                $filesystem = Filesystem::mount(Path::of($path));
+                $filesystem = Filesystem::mount(Path::of($path))->unwrap();
 
                 $this->assertInstanceOf(
                     SideEffect::class,
@@ -396,7 +398,7 @@ class FilesystemTest extends TestCase
                 $path = \sys_get_temp_dir().'/innmind/filesystem/';
                 (new FS)->remove($path);
 
-                $filesystem = Filesystem::mount(Path::of($path));
+                $filesystem = Filesystem::mount(Path::of($path))->unwrap();
 
                 $this->assertInstanceOf(
                     SideEffect::class,
@@ -431,7 +433,7 @@ class FilesystemTest extends TestCase
                 $path = \sys_get_temp_dir().'/innmind/filesystem/';
                 (new FS)->remove($path);
 
-                $filesystem = Filesystem::mount(Path::of($path));
+                $filesystem = Filesystem::mount(Path::of($path))->unwrap();
 
                 $this->assertInstanceOf(
                     SideEffect::class,
@@ -457,7 +459,7 @@ class FilesystemTest extends TestCase
         (new FS)->dumpFile($path.'foo', 'bar');
         \symlink($path.'foo', $path.'bar');
 
-        $filesystem = Filesystem::mount(Path::of($path));
+        $filesystem = Filesystem::mount(Path::of($path))->unwrap();
 
         $this->expectException(LinksAreNotSupported::class);
         $this->expectExceptionMessage($path.'bar');
@@ -472,7 +474,7 @@ class FilesystemTest extends TestCase
         (new FS)->dumpFile($path.'foo', 'bar');
         \symlink($path.'foo', $path.'bar');
 
-        $filesystem = Filesystem::mount(Path::of($path));
+        $filesystem = Filesystem::mount(Path::of($path))->unwrap();
 
         $this->expectException(LinksAreNotSupported::class);
         $this->expectExceptionMessage($path.'bar');
@@ -490,7 +492,7 @@ class FilesystemTest extends TestCase
                 (new FS)->mkdir($path);
                 \file_put_contents($path.$name, 'bar');
 
-                $filesystem = Filesystem::mount(Path::of($path));
+                $filesystem = Filesystem::mount(Path::of($path))->unwrap();
 
                 $all = $filesystem->root()->all()->toList();
                 $this->assertCount(1, $all);