+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Requires PHP `8.4`
 - Links are not silently ignored when listing files/directories. An error is still returned when trying to remove a link.
+- `Innmind\Filesystem\Adapter\Filesystem::mount()` now returns an `Innmind\Immutable\Attempt`
 
 ## 8.1.0 - 2025-05-09